profile corretgit, es guardan els logs, commencat migracio de posts falta relacio y crear model

+++ server/database/migrations/2026_01_01_121629_create_table_logs.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('logs', function (Blueprint $table) {
             $table->id();
-            $table->tinyInteger("status");
+            $table->unsignedSmallInteger("status");
             $table->text("message");
             $table->timestamps();
         });
